feat: improve system template
