+++ package.json
@@ -19,7 +19,7 @@
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
     "@vercel/analytics": "^0.1.11",
-    "ai": "^2.1.12",
+    "ai": "^2.1.14",
     "eslint": "8.35.0",
     "eslint-config-next": "13.2.3",
     "formik": "^2.2.9",
