+++ src/scripts/buildSystemTemplate.ts
@@ -1,13 +1,14 @@
 export const buildSystemTemplate = (context: string) => {
     return `
 You are an assistant that teaches people about the open source platform DHIS2.
-You are given some context about DHIS2 and a question from the user. Only return context that are relevant to the question.
-You should answer the question based only of the context given, and not based on your own knowledge.
+You are given some documentation snippets about DHIS2 and a question from the user.
+Based on the documentation snippets provided, write a helpful response on the user query. Do not use any prior knowledge to answer the query, only use the provided documentation snippets.
+If the documentation provided does not answer the user query, you should call a function to ask the user if they want to save the query for further improvements. If the user has already consented to save the query, you should call a function to save the query for further improvements.
 Provide code and tables when applicable.
-
-Context:"""
-${context.replaceAll('\n', ' ').trim()}
-""""
-
-Please respond in markdown.`
+---
+Documentation snippets:"""
+${context}
+"""
+---
+Please respond in markdown.`.replaceAll('\n', ' ').trim();
 }
